Add challenge evaluator and prediction generator scripts
Added Prediction_Validation/challenge_evaluator.py for local evaluation of object detection predictions and WorkStation_MoE/Prediction_Json_Generator.py for generating prediction JSON files. Updated .gitignore to exclude new prediction and output files. Fixed import paths in MMMMoE.py, changed label dtype in IRJsonDataset.py to int32, updated checkpoint path in Validation.py, and removed WorkStation_Rulebase/main.py.

+++ WorkStation_MoE_Month_Hour/IRJsonDataset.py
@@ -228,7 +228,7 @@ class IRJsonDataset(Dataset):
                 labels_np = np.zeros((0, 1), dtype=np.float32)
             else:
                 bboxes_np = np.asarray(boxes_list, dtype=np.float32).reshape(-1, 4)
-                labels_np = np.asarray(labels_list, dtype=np.float32).reshape(-1, 1)
+                labels_np = np.asarray(labels_list, dtype=np.int32).reshape(-1, 1)
 
             segments = np.zeros((0, 0, 2), dtype=np.float32)
